medical_migration_1 /migration/requirements.txt
pymongo==4.6.1
pandas==2.2.1
python-dotenv==1.0.1
requests==2.31.0
